Update the discussion of platform security to reflect modern macOS and Windows capabilities
The description of what security features the keychain APIs enable isn't quite up to date. It's possible to get security with this library as long as specific criteria are met. This PR updates the discussion to describe how to ensure keychain items are properly protected. 

Disclosure: getting keychain security does require your app to be packaged in specific ways. At this time the only way I know of to package JVM apps in a secure manner is with Conveyor, a tool linked to in the discussion. My company makes Conveyor. I'm updating this README because we use this library in the app itself, and because the info might be useful to other JVM users. Conveyor is a commercial tool but free for open source projects.